+++ changes.md
@@ -4,6 +4,7 @@
 ### 4.0.1+ (Build: 20220512)
 - Updated Moodle core
 - Updated 'completion progress' block to 2022042000
+- Updated 'etask' course format to 2022041901
 
 ### 4.0+ (Build: 20220429)
 - Initial checkout of Moodle 4.0.0